requirements.txt
attrs==20.3.0
aws-cdk.assets==1.118.0
aws-cdk.aws-cloud9==1.118.0
aws-cdk.aws-cloudwatch==1.118.0
aws-cdk.aws-codecommit==1.118.0
aws-cdk.aws-ec2==1.118.0
aws-cdk.aws-events==1.118.0
aws-cdk.aws-iam==1.118.0
aws-cdk.aws-kms==1.118.0
aws-cdk.aws-logs==1.118.0
aws-cdk.aws-s3==1.118.0
aws-cdk.aws-s3-assets==1.118.0
aws-cdk.aws-ssm==1.118.0
aws-cdk.cloud-assembly-schema==1.118.0
aws-cdk.core==1.118.0
aws-cdk.cx-api==1.118.0
aws-cdk.region-info==1.118.0
cattrs==1.6.0
constructs==3.3.123
jsii==1.32.0
publication==0.0.3
python-dateutil==2.8.2
six==1.16.0
typing-extensions==3.10.0.0
